Corrected some comments and modified makefiles to take into account new registry.c file.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@8439 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/library/wel/clib/msgboxpa.c
@@ -28,7 +28,7 @@ cwel_msgboxparams_set (LPMSGBOXPARAMS item, HWND wd, HINSTANCE hinst, LPCSTR txt
 	item -> lpszCaption = ttl;
 	item -> dwStyle = s;
 	item -> lpszIcon = ico;
-	item -> dwContextHelpId = NULL;
+	item -> dwContextHelpId = (DWORD)NULL;
 	item -> lpfnMsgBoxCallback = NULL;
 	item -> dwLanguageId = langid;
 }
